requirements.txt
# Core dependencies
streamlit>=1.31.0
langchain>=0.1.0
chromadb>=0.4.22
python-dotenv>=1.0.0

# Document Processing
pypdf>=3.17.1
pdfminer.six>=20221105
python-docx>=1.0.0
reportlab>=4.0.0

# NLP & Data Processing
spacy>=3.7.2
nltk>=3.8.1
beautifulsoup4>=4.12.2
lxml[html_clean]>=5.1.0
requests>=2.31.0

# Machine Learning
transformers>=4.36.2
torch>=2.1.0

# Utilities
python-dateutil>=2.8.2

# Development Tools
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.1